+++ ninja-core/src/main/java/ninja/RouteBuilderImpl.java
@@ -28,8 +28,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.inject.Injector;
+import ninja.utils.MethodReference;
 
-class RouteBuilderImpl implements RouteBuilder {
+public class RouteBuilderImpl implements RouteBuilder {
 
     private static final Logger log = LoggerFactory
             .getLogger(RouteBuilder.class);
@@ -80,6 +81,11 @@ class RouteBuilderImpl implements RouteBuilder {
         this.controller = controller;
         this.controllerMethod = verifyThatControllerAndMethodExists(controller,
                 controllerMethod);
+    }
+
+    @Override
+    public void with(MethodReference controllerMethodRef) {
+        with(controllerMethodRef.getDeclaringClass(), controllerMethodRef.getMethodName());
     }
 
     @Override