+++ ninja-core/src/main/java/ninja/cache/NinjaCache.java
@@ -20,14 +20,13 @@ import java.io.NotSerializableException;
 import java.io.Serializable;
 import java.util.Map;
 
-import net.sf.ehcache.CacheException;
 import ninja.utils.TimeUtil;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
 /**
- * A convenience class to access the underlaying cache implementation.
+ * A convenience class to access the underlying cache implementation.
  * 
  * Makes getting and setting of objects a lot simpler.
  *