+++ ninja-core/src/test/java/ninja/bodyparser/BodyParserEngineManagerImplTest.java
@@ -64,6 +64,10 @@ public class BodyParserEngineManagerImplTest {
                 bind(Lang.class).to(LangImpl.class);
                 bind(Router.class).to(RouterImpl.class);
                 Multibinder.newSetBinder(binder(), ParamParser.class);
+
+                bind(BodyParserEnginePost.class);
+                bind(BodyParserEngineJson.class);
+                bind(BodyParserEngineXml.class);
                 
                 bind(NinjaProperties.class).toInstance(new NinjaPropertiesImpl(NinjaMode.test));