+++ ninja-core/src/test/java/ninja/standalone/StandaloneHelperTest.java
@@ -17,12 +17,20 @@
 package ninja.standalone;
 
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
 import javax.net.ssl.SSLContext;
+import ninja.utils.ForwardingServiceLoader;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.fail;
 import org.junit.Test;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class StandaloneHelperTest {
 
@@ -53,6 +61,42 @@ public class StandaloneHelperTest {
             = StandaloneHelper.createSSLContext(keystoreUri, keystorePassword, truststoreUri, truststorePassword);
         
         assertThat(sslContext, is(not(nullValue())));
+    }
+    
+    @Test
+    public void resolveStandaloneClass() throws Exception {
+        Class<? extends Standalone> resolvedStandaloneClass;
+        
+        try {
+            // try to resolve system property
+            StandaloneHelper.resolveStandaloneClass("sys_prop_does_not_exist", null, null);
+            fail();
+        } catch (Exception e) {
+            assertThat(e.getMessage(), containsString("sys_prop_does_not_exist"));
+        }
+        
+        // try to find using META-INF/services
+        // mock the service loader
+        ForwardingServiceLoader<Standalone> serviceLoader
+            = mock(ForwardingServiceLoader.class);
+
+        List<Standalone> standalones = new ArrayList<>();
+        standalones.add(new MockStandalone());
+        when(serviceLoader.iterator()).thenReturn(standalones.iterator());
+
+        resolvedStandaloneClass = StandaloneHelper.resolveStandaloneClass(null, serviceLoader, "default_does_not_exist");
+       
+        assertThat(resolvedStandaloneClass.newInstance(), instanceOf(MockStandalone.class));
+        
+        try {
+            // try to resolve using default
+            when(serviceLoader.iterator()).thenReturn(new ArrayList<Standalone>().iterator());
+            StandaloneHelper.resolveStandaloneClass(null, serviceLoader, "default_does_not_exist");
+            fail();
+        } catch (Exception e) {
+            assertThat(e.getMessage(), containsString("default_does_not_exist"));
+        }
+        
     }
     
 }